# Conflicts:
#	CMakeLists.txt
#	Common/core/def_version.h
#	Common/game/main_game_file.cpp
#	Common/game/main_game_file.h
#	Common/gui/guilistbox.cpp
#	Common/gui/guitextbox.cpp
#	Editor/AGS.Editor/AGSEditor.cs
#	Editor/AGS.Editor/ColorThemes/Controls/ComboBoxCustom.cs
#	Editor/AGS.Editor/DataFileWriter.cs
#	Editor/AGS.Editor/app.manifest
#	Editor/AGS.Native/agsnative.cpp
#	Editor/AGS.Types/Properties/AssemblyInfo.cs
#	Editor/AGS.Types/Settings.cs
#	Engine/ac/display.cpp
#	Engine/ac/draw.cpp
#	Engine/ac/draw.h
#	Engine/ac/drawingsurface.cpp
#	Engine/ac/global_display.cpp
#	Engine/ac/global_drawingsurface.cpp
#	Engine/ac/global_gui.cpp
#	Engine/ac/global_object.cpp
#	Engine/ac/room.cpp
#	Engine/game/savegame_v321.cpp
#	Engine/gui/gui_engine.cpp
#	Engine/main/main.cpp
#	Engine/main/quit.cpp
#	Engine/platform/android/acpland.cpp
#	Engine/platform/base/agsplatformdriver.h
#	Engine/platform/windows/acplwin.cpp
#	Engine/script/cc_instance.cpp
#	Engine/script/cc_instance.h
#	Engine/script/script.cpp
#	version.json

+++ Engine/ac/global_string.cpp
@@ -21,7 +21,7 @@
 #include "ac/string.h"
 #include "util/string_compat.h"
 
-extern int MAXSTRLEN;
+extern size_t MAXSTRLEN;
 
 // CLNUP probably to remove
 /*int StrGetCharAt (const char *strin, int posn) {
